requirements.txt
python-telegram-bot==20.4
pymongo==4.8.0
python-dotenv==1.0.1
certifi